pyproject.toml
[tool.poetry]
name = "aanntwin"
version = "0.1.0"
description = ""
authors = ["Dimitar Dimitrov <dimitrov.d@northeastern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
# torch = "^2.1.0"
# torchvision = "^0.16.0"
gitpython = "^3.1.40"
numpy = "^1.24"

[tool.poetry.group.dev.dependencies]
ipython = "^8.12.3"
reorder-python-imports = "^3.12.0"
black = "^23.9.1"
pylint = "^3.0.1"
mypy = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
